[SYCL][Sema] Prevent Ctor/Dtor from being called during SYCL kernel rebuild
User can declare their type as device copyable even
if they do not meet the criterias. In some cases,
this triggers calls to undefined, deleted or not allowed on device functions.

Per SYCL rules, a SYCL kernel object must be device copiable
which implies the structure can be initialized using memcpy
and have a trivial destructor. Plus the spec doesn't guarantee
that the constructor/destructor should get called at all as it allows the functor
to be reused by multiple WI.

This patch exploit that implementation defined behavior by preventing the constructor/destructor
of being called by wrapping it into a union.

In order to prevent this, the SYCL kernel is wrapped into a union and
instead of using initializer list,
the SYCL kernel clone is init using assignment for scalars, memcpy for non special types
and Ctor/__init for SYCL special types.

Signed-off-by: Victor Lomuller <victor@codeplay.com>